Fix two bugs exposed by automated testing
- It would throw errors when attempting to remove a component from a
GameObject, because it was calling GetComponent on the prop type rather
than the component type. To fix this, needed to add the component type
to the CompElem constructor, and add a new RemoveComponent function to
CompElem.
- It was not properly updating the parentKey on the PopulatedNodeElem
when changing a node's position in the scene graph. This resulted in
nodes thinking they're in a different position than they really are,
which potentially breaks future moves.