Fix a crash when adding new nodes
When the new render adds a node that wasn't in the previous render, it
would crash. This was because it was executing the creation queue before
the update queue, so dictionary of new populated nodes did not yet
contain all the existing nodes at the point where it's making new nodes,
so the attempt to look up the parent node failed.

Fixed by simply reordering the render to run the update queue before the
creation queue.

Also fix a bug which was exposed by the test update for this, which is
that newly created nodes were not having their parentKey set properly,
if they weren't a root node.