Substantially improve TransformComponent
- Add nullable lookAt property. If present, the rotation property is
ignored, and this is used instead, to rotate the node to point toward
the point specified. The up vector is asssumed to me positive Y.
- Made localScale optional. The default value for a Vector3 is all zeros
which is the wrong default for scale. So it was necessary to distinguish
explicitly when localScale isn't set so a sane default (all ones) could
be used instead.
- Always set the current transform values, instead of diffing against
the previous state. Since the actual transform can be altered by the
parent GameObject's transform changing, comparing against the props from
the previous frame is insufficient to know if it needs to be set again
this frame. Also, setting the transform every frame is a very standard
and optimized thing to do in Unity, so this isn't a significant
performance issue.