Add the concept of mesh/material constructor
The MeshFilter and MeshRenderer needed to have a mesh or material
specified in the props. This works fine when those are pre-loaded assets
that can just be passed in from some global asset store or something.
However, when the mesh or material is composed on the fly, this can be
problematic, as you don't typically want to actually recreate them every
single frame.

So, a new property was added to MeshFilter and MeshRenderer, called
meshConstructor and materialConstructor, respectively. These each take
a lambda, which returns a mesh or a material. This lambda is only called
on the first frame that these properties are set. This allows you to
programatically build a mesh or material, without having it be redone
every frame. The catch is that it can't diff the lambda properly, as
every frame a new lambda is created, so it will always compare as
unequal even when it's an identical function. You could get around this
by passing in a static function, but that would mean you could provide
no arguments to it. Which only makes sense if it's a static mesh, the
same every time, which means there'd be no reason to be using the
constructor at all. So, as long as the constructor was set last frame
and is still set this frame, it assumes it hasn't changed, and does not
run it again. If you need to update the mesh, you need to unset the
lambda for a frame, and reset it the next frame.

A better solution which encodes a property with a constructor in a
standardized way, and moves it's arguments into a diff-able property
structure to allow better updating, is being considered, but it's a bit
complicated to make that work right and this will handle most cases for
now.