Simplify the API by removing prop structs
The React convention of using prop objects makes a lot of sense in
JavaScript, where all objects are just dictionaries that can be iterated
and compared. Doing the same with a struct in C# would require heavy use
of reflection, which would not be performant, so trying to hold to that
convention just complicated the UI with little real benefit.

Previously, a UReact component was the combination of a Unity component
type, a render function and a prop struct instance. This relied heavily
on following conventions. Now, a UReact component is defined by the
UReact.Component interface. This has a Render function on it, which is
much like the old Render function, but rather than being passed the old
props and the current props, it's just passed the old component. The
component instance is expected to take any properties it needs in it's
constructor and store them directly. The Render function will tend to be
structred much as before, except instead of comparing oldProps to props,
it's comparing oldComp to itself.

The interface also defines the GetManagedBehaviourTypes function, which
returns a Type[], which is expected to be populated with all the Unity
component types this UReact component manages. This is used by the
cleanup code to know what Unity components to remove from the GameObject
when this UReact component goes away.

The API for nodes is unchanged, but the convention is. Now, instead of
the node's New function taking a prop struct, it just takes all the
properties directly as function arguments.

By convention, both the node New function, and the constructors for
UReact components, should be called entirely with named parameters,
which means the calling convention looks very similar. The upside of
using function arguments instead of a struct is that it can now be
explicit which properties are required and which are optional, and be
explicit about the default value used for optional properties, and have
the manditory properties actually enforced by the compiler. The old
struct construction approach would allow any property to be optional,
and would simply fill in the general default value for that data type.
Which wasn't a great behavior.

Also updated the tests and documentation to match.

Finally, added two minor features to UReact:
- Renderer.Clear will clear out the actual GameObject hierarchy.
Should always call this when you're done with a given renderer, so it
doesn't leave it's scene graph behind.
- Renderer.IsInSceneGraph is passed a GameObject, and will return true
if that GameObject is in the portion of the scene graph managed by this
renderer.